Aggiunti commenti

+++ Recognition.py
@@ -4,6 +4,7 @@ import numpy as np
 import LBP
 import math
 import ast
+import os.path
 from matplotlib import pyplot as plt
 
 # ritorna i dati del paziente cf e del delgato
@@ -251,6 +252,7 @@ def evaluationIdentificationAsMultiVer():
     print("FRR:", FRR, countTG)
     print("FAR:", FAR, countTI)
 
+# restituisce il numero di False Accept, False Reject, True Genuine e True Impostor
 def delegatesMatch(hist_data, target, gallery_target, cf_list, users, gallery_thresholds, histogram_gallery_data):
 
     # inizializzazione delle variabili
@@ -262,36 +264,45 @@ def delegatesMatch(hist_data, target, gallery_target, cf_list, users, gallery_th
 
     # per ogni istogramma
     for i in range(len(hist_data)):
-        probe_target = target[i]
-        hist_probe = hist_data[i]
+        probe_target = target[i]        #identità a cui appartiene l'istogramma
+        hist_probe = hist_data[i]       #istogramma
+        # per ogni utente della gallery
         for j in range(len(galley_users)):
-                cf_user = galley_users[j]
+                cf_user = galley_users[j]       #codice fiscale dell'utente
                 index = cf_list.tolist().index(cf_user)
-                user = users.iloc[index]
-                delegati = ast.literal_eval(user["Delegati"])
-
-                # se l'utente ha dei delegati
+                user = users.iloc[index]        #informazioni dell'utente
+                delegati = ast.literal_eval(user["Delegati"])       #lista dei delegati dell'utente
+                # se l'utente ha dei delegati e se l'identità dell'utente non corrisponde a quella del probe
                 if len(delegati) > 0 and cf_user != probe_target:
+                    # se il probe è un delegato, allora è un True Genuine
                     if probe_target in delegati:
                         countTG += 1
+                    # altrimenti è un True Impostor
                     else:
                         countTI += 1
                     accepted = False
+                    # per ogni delegato
                     for t in delegati:
+                        # restituisce la migliore corrispondenza tra hist_probe e gli istogrammi del delegato nella galleria
                         val = topMatch(t, gallery_target, histogram_gallery_data, hist_probe)
                         index_threshold = cf_list.tolist().index(t)
+                        # se la similarità supera il threshold adattivo, allora viene riconosciuto come uno dei delegati
                         if val >= gallery_thresholds[index_threshold]:
                             accepted = True
                             continue
+                    # se il probe viene riconosciuto ma non è un delegato, incremente False Accept
                     if accepted and probe_target not in delegati:
                         fa += 1
+                    # se il probe non viene riconosciuto nonostante sia un delegato, incremente False Reject
                     elif not accepted and probe_target in delegati:
                         fr += 1
         print(probe_target, fr, countTG, fa, countTI)
     return fa, fr, countTG, countTI
 
+# mostra i False Rejection Rate, i False Acceptance Rate, l'Equal Error Rate e la ROC curve nel caso in cui si adottano i threshold fissi
 def evaluationVerification():
 
+    # caricamento dei vari dataset
     pg_target = np.load("npy_db/pg_target.npy")
     pn_target = np.load("npy_db/pn_target.npy")
     gallery_target = np.load("npy_db/gallery_target.npy")
@@ -300,30 +311,35 @@ def evaluationVerification():
     histogram_pn_data = np.load("npy_db/histogram_pn_data.npy")
     galley_identities = list(dict.fromkeys(gallery_target))
 
-    similarity_matrix = np.zeros((len(histogram_pg_data)+len(histogram_pn_data), len(galley_identities)))
-
-    for x in range(len(histogram_pg_data)+len(histogram_pn_data)):
-        # seleziona l'istogramma del probe
-        if x < len(histogram_pg_data):
-            hist_probe = histogram_pg_data[x]
-        else:
-            hist_probe = histogram_pn_data[x-len(histogram_pg_data)]
-
-        # Sign for each identity the max similarity against probe
-        for y in range(len(galley_identities)):
-            claimed_identity = galley_identities[y]
-            similarity_matrix[x][y] = topMatch(claimed_identity, gallery_target, histogram_gallery_data, hist_probe)
-
-    np.save("similarity_matrix.npy", similarity_matrix)
-
-    similarity_matrix = np.load("similarity_matrix.npy")
-
+    # se non esiste un file contenente la similarity matrix, allora viene creato
+    if not os.path.exists("similarity_matrix.npy"):
+        # creazione della similarity matrix
+        similarity_matrix = np.zeros((len(histogram_pg_data)+len(histogram_pn_data), len(galley_identities)))
+        # per ogni istogramma ottenuto dai probe
+        for x in range(len(histogram_pg_data)+len(histogram_pn_data)):
+            # seleziona l'istogramma del probe gallery
+            if x < len(histogram_pg_data):
+                hist_probe = histogram_pg_data[x]
+            # seleziona l'istogramma del probe non appartenete alla gallery
+            else:
+                hist_probe = histogram_pn_data[x-len(histogram_pg_data)]
+            # segna per ogni identità la massima somiglianza con il probe
+            for y in range(len(galley_identities)):
+                claimed_identity = galley_identities[y]
+                similarity_matrix[x][y] = topMatch(claimed_identity, gallery_target, histogram_gallery_data, hist_probe)
+        # salva la similarity matrix
+        np.save("similarity_matrix.npy", similarity_matrix)
+    # se il file esiste, la matrice è stata già creata e quindi viene caricata
+    else:
+        similarity_matrix = np.load("similarity_matrix.npy")
+    # calcolo dei True Genuine e dei True Impostors
     tg = len(pg_target)
     ti = len(pg_target)*(len(galley_identities)-1) + len(pn_target)*len(galley_identities)
     print("TG:", tg)
     print("TI:", ti)
     print()
 
+    # inizializzazione delle variabili che conterranno i vari rate e thresholds
     frr = []
     far = []
     gar = []
@@ -333,7 +349,7 @@ def evaluationVerification():
     # per ogni threshold
     for threshold in np.arange(0.0, 1.01, 0.01):
 
-        # si fissa il threshold
+        # si fissa la soglia
         th = np.round(threshold, 2)
 
         # si resettano le variabili
@@ -342,35 +358,35 @@ def evaluationVerification():
         fr = 0
         gr = 0
 
-        # per ogni probe prendiamo l'identità reale e il suo istogramma
+        # per ogni probe prendiamo l'identità reale e il suo istogramma (ogni riga corrisponde ad un probe)
         for x in range(similarity_matrix.shape[0]):
             if x < len(histogram_pg_data):
                 real_identity = pg_target[x]
             else:
                 real_identity = pn_target[x-len(histogram_pg_data)]
 
-            # We claim every identity enrolled in the gallery
+            # per ogni identità iscritta alla gallery (ogni colonna corrisponde ad un utente)
             for y in range(similarity_matrix.shape[1]):
                 claim_identity = galley_identities[y]
                 similarity = similarity_matrix[x][y]
 
-                # If the max similarity is above the threshold:
+                # se la similairty massima è al di sopra della soglia
                 if similarity >= th:
-                    # If the claimed identity is truw, we increase the GA
+                    # e se l'identità dichiarata è vera, aumenta Genuine Accept
                     if real_identity == claim_identity:
                         ga += 1
-                    # Otherwise we increase the FA
+                    # altrimenti incrementa  False Accept
                     else:
                         fa += 1
-                # If the max similarity is below the threshold:
+                # se la similairty massima è inferiore alla soglia
                 else:
-                    # If the claimed identity is truw, we increase the FR
+                    # e se l'identità dichiarata è vera, aumenta False Reject
                     if real_identity == claim_identity:
                         fr += 1
-                    # Otherwise we increase the GR
+                    # altrimenti incrementa Genuine Reject
                     else:
                         gr += 1
-
+        # aggiungiamo i rate relativi al threshold th alle varie liste
         gar.append(ga/tg)
         far.append(fa/ti)
         frr.append(fr/tg)
@@ -383,18 +399,19 @@ def evaluationVerification():
     print("GRR:", grr)
     print()
 
+    # cacolo dell'Equal Error Rate
     eer_1 = np.array(far)[np.nanargmin(np.absolute((np.array(frr) - np.array(far))))]
     eer_2 = np.array(frr)[np.nanargmin(np.absolute((np.array(frr) - np.array(far))))]
     eer = (eer_1 + eer_2) / 2
-
     print("EER:", eer)
-
+    # cacolo del threshold dell'Equal Error Rate
     eer_threshold = np.array(thresholds)[np.nanargmin(np.absolute((np.array(frr) - np.array(far))))]
-
     print("EER Threshold:", eer_threshold)
     print()
+
     print("Thresholds:", thresholds)
 
+    # traccia la ROC curve
     plt.plot(far, gar)
     plt.plot([0, 1], [0, 1], 'k--')
     plt.ylabel("Genuine Acceptance Rate")
@@ -448,22 +465,4 @@ def verificationAdaptive():
     return
 
 if __name__ == '__main__':
-    #verificationFRR()
-    #verificationFAR()
     evaluationIdentificationAsMultiVer()
-    #verificationAdaptive()
-    #evaluationVerification()
-
-    #frr = [0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.007142857142857143, 0.007142857142857143, 0.007142857142857143, 0.007142857142857143, 0.007142857142857143, 0.02142857142857143, 0.02142857142857143, 0.02142857142857143, 0.02142857142857143, 0.02857142857142857, 0.02857142857142857, 0.04285714285714286, 0.05, 0.06428571428571428, 0.07142857142857142, 0.09285714285714286, 0.12857142857142856, 0.15, 0.17142857142857143, 0.17857142857142858, 0.20714285714285716, 0.24285714285714285, 0.2785714285714286, 0.3, 0.34285714285714286, 0.38571428571428573, 0.45, 0.5357142857142857, 0.5785714285714286, 0.6285714285714286, 0.6714285714285714, 0.75, 0.8, 0.8428571428571429, 0.8785714285714286, 0.9214285714285714, 0.9428571428571428, 0.9571428571428572, 0.9857142857142858, 0.9857142857142858, 0.9928571428571429, 0.9928571428571429, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0]
-    #far = [1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 0.9997198879551821, 0.9997198879551821, 0.9991596638655462, 0.9984593837535014, 0.9974789915966387, 0.9957983193277311, 0.9928571428571429, 0.9882352941176471, 0.9831932773109243, 0.9746498599439776, 0.9644257703081233, 0.9511204481792717, 0.9329131652661065, 0.9119047619047619, 0.8813725490196078, 0.846078431372549, 0.8081232492997199, 0.7675070028011205, 0.7177871148459384, 0.6668067226890756, 0.6095238095238096, 0.5497198879551821, 0.49089635854341734, 0.426750700280112, 0.3680672268907563, 0.315266106442577, 0.26554621848739496, 0.22198879551820727, 0.18207282913165265, 0.14523809523809525, 0.11708683473389356, 0.09243697478991597, 0.07226890756302522, 0.05742296918767507, 0.044677871148459385, 0.033473389355742296, 0.025070028011204483, 0.018067226890756304, 0.012324929971988795, 0.008963585434173669, 0.0056022408963585435, 0.0036414565826330533, 0.0021008403361344537, 0.0016806722689075631, 0.0014005602240896359, 0.0004201680672268908, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0]
-    #thresholds = [0.0, 0.01, 0.02, 0.03, 0.04, 0.05, 0.06, 0.07, 0.08, 0.09, 0.1, 0.11, 0.12, 0.13, 0.14, 0.15, 0.16, 0.17, 0.18, 0.19, 0.2, 0.21, 0.22, 0.23, 0.24, 0.25, 0.26, 0.27, 0.28, 0.29, 0.3, 0.31, 0.32, 0.33, 0.34, 0.35, 0.36, 0.37, 0.38, 0.39, 0.4, 0.41, 0.42, 0.43, 0.44, 0.45, 0.46, 0.47, 0.48, 0.49, 0.5, 0.51, 0.52, 0.53, 0.54, 0.55, 0.56, 0.57, 0.58, 0.59, 0.6, 0.61, 0.62, 0.63, 0.64, 0.65, 0.66, 0.67, 0.68, 0.69, 0.7, 0.71, 0.72, 0.73, 0.74, 0.75, 0.76, 0.77, 0.78, 0.79, 0.8, 0.81, 0.82, 0.83, 0.84, 0.85, 0.86, 0.87, 0.88, 0.89, 0.9, 0.91, 0.92, 0.93, 0.94, 0.95, 0.96, 0.97, 0.98, 0.99, 1.0]
-
-    # index = thresholds.index(0.62)
-    # print("Valori ottenuti utilizzando il eer_threshold = 0.62")
-    # print("FRR:", frr[index])
-    # print("FAR:", far[index])
-    # index = far.index(0.0036414565826330533)
-    # print("Valori ottenuti utilizzando un threshold =", thresholds[index], "che ci dia un far simile a quello ottenuto con i threshold adattivi")
-    # print("FRR:", frr[index])
-    # print("FAR:", far[index])
-
